Small fixes to type warnings, unused local variables, and arguments that obscure class members.

+++ src/core/InputHandler.cpp
@@ -94,7 +94,7 @@ namespace vulpes {
         
         MouseScroll += static_cast<float>(y_offset);
         ImGuiIO& io = ImGui::GetIO();
-        io.MouseWheel = y_offset;
+        io.MouseWheel = static_cast<float>(y_offset);
 
 	}
